move zip to module and clean up options
